services/dev_agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pyautogui>=0.9.54
Pillow>=10.0.0
httpx>=0.27.0
